+++ NEWS.md
@@ -1,6 +1,8 @@
 # energy 1.7-9
 
-*  bug fix in normal.test
+  *  bug fix in normal.test
+  *  pdcor.test now gives p.value=NA in degenerate case
+  *  DCOR is deprecated; use dcor or dcov
 
 # energy 1.7-8